+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/vm/VM.java
@@ -56,6 +56,7 @@ import java.util.Properties;
 import java.util.StringJoiner;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicLongFieldUpdater;
 import java.util.function.IntFunction;
 
 import org.graalvm.collections.EconomicMap;
@@ -129,7 +130,6 @@ import com.oracle.truffle.espresso.nodes.EspressoFrame;
 import com.oracle.truffle.espresso.nodes.EspressoRootNode;
 import com.oracle.truffle.espresso.nodes.interop.ToEspressoNode;
 import com.oracle.truffle.espresso.nodes.interop.ToEspressoNodeGen;
-import com.oracle.truffle.espresso.overlay.ReferenceSupport;
 import com.oracle.truffle.espresso.ref.EspressoReference;
 import com.oracle.truffle.espresso.runtime.Attribute;
 import com.oracle.truffle.espresso.runtime.Classpath;
@@ -155,8 +155,6 @@ import com.oracle.truffle.espresso.vm.structs.JavaVMAttachArgs;
 import com.oracle.truffle.espresso.vm.structs.JdkVersionInfo;
 import com.oracle.truffle.espresso.vm.structs.Structs;
 import com.oracle.truffle.espresso.vm.structs.StructsAccess;
-
-import sun.misc.Unsafe;
 
 /**
  * Espresso implementation of the VM interface (libjvm).
@@ -592,16 +590,26 @@ public final class VM extends NativeEnv {
         return Double.isNaN(d);
     }
 
+    private static final class LongCASProbe {
+        static final boolean HOST_SUPPORTS_LONG_CAS;
+        @SuppressWarnings("unused") volatile long foo;
+        static {
+            AtomicLongFieldUpdater<LongCASProbe> updater = AtomicLongFieldUpdater.newUpdater(LongCASProbe.class, "foo");
+            String updaterName = updater.getClass().getSimpleName();
+            if ("CASUpdater".equals(updaterName)) {
+                HOST_SUPPORTS_LONG_CAS = true;
+            } else if ("LockedUpdater".equals(updaterName)) {
+                HOST_SUPPORTS_LONG_CAS = false;
+            } else {
+                throw EspressoError.shouldNotReachHere("Unknown host long updater: " + updaterName);
+            }
+        }
+    }
+
     @VmImpl
     @TruffleBoundary
     public static boolean JVM_SupportsCX8() {
-        try {
-            java.lang.reflect.Field field = AtomicLong.class.getDeclaredField("VM_SUPPORTS_LONG_CAS");
-            Unsafe unsafe = UnsafeAccess.get();
-            return unsafe.getBoolean(unsafe.staticFieldBase(field), unsafe.staticFieldOffset(field));
-        } catch (NoSuchFieldException e) {
-            throw EspressoError.shouldNotReachHere(e);
-        }
+        return LongCASProbe.HOST_SUPPORTS_LONG_CAS;
     }
 
     @VmImpl(isJni = true)
@@ -3628,7 +3636,7 @@ public final class VM extends NativeEnv {
         }
         assert host instanceof Reference : host;
         // Call host's refersTo. Not available in 8 or 11.
-        return ReferenceSupport.phantomReferenceRefersTo((Reference) host, object);
+        return ((Reference<StaticObject>) host).refersTo(object);
     }
 
     @SuppressWarnings({"rawtypes", "unchecked"})
@@ -3651,7 +3659,7 @@ public final class VM extends NativeEnv {
             }
             assert host instanceof Reference : host;
             // Call host's refersTo. Not available in 8 or 11.
-            return ReferenceSupport.referenceRefersTo((Reference) host, object);
+            return ((Reference<StaticObject>) host).refersTo(object);
         } else {
             StaticObject referent = (StaticObject) meta.java_lang_ref_Reference_referent.get(ref);
             return InterpreterToVM.referenceIdentityEqual(referent, object, language);