+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -50,6 +50,7 @@ import java.io.Reader;
 import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.Executable;
 import java.lang.reflect.Type;
+import java.math.BigInteger;
 import java.net.URI;
 import java.net.URL;
 import java.nio.ByteOrder;
@@ -237,6 +238,8 @@ public abstract class AbstractPolyglotImpl {
         public abstract boolean isIteratorAccessible(HostAccess access);
 
         public abstract boolean isMapAccessible(HostAccess access);
+
+        public abstract boolean isBigIntegerAccessibleAsNumber(HostAccess access);
 
         public abstract boolean allowsPublicAccess(HostAccess hostAccess);
 
@@ -959,6 +962,12 @@ public abstract class AbstractPolyglotImpl {
         }
 
         public abstract long asLong(Object context, Object receiver);
+
+        public boolean fitsInBigInteger(Object context, Object receiver) {
+            return false;
+        }
+
+        public abstract BigInteger asBigInteger(Object context, Object receiver);
 
         public boolean fitsInDouble(Object context, Object receiver) {
             return false;