+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/thread/Target_java_lang_Thread.java
@@ -33,7 +33,6 @@ import java.util.Objects;
 import java.util.concurrent.ThreadFactory;
 import java.util.function.BooleanSupplier;
 
-import com.oracle.svm.util.ReflectionUtil;
 import org.graalvm.compiler.api.directives.GraalDirectives;
 import org.graalvm.compiler.replacements.ReplacementsUtil;
 import org.graalvm.compiler.serviceprovider.JavaVersionUtil;
@@ -52,6 +51,7 @@ import com.oracle.svm.core.annotate.RecomputeFieldValue;
 import com.oracle.svm.core.annotate.Substitute;
 import com.oracle.svm.core.annotate.TargetClass;
 import com.oracle.svm.core.annotate.TargetElement;
+import com.oracle.svm.core.deopt.DeoptimizationSupport;
 import com.oracle.svm.core.jdk.ContinuationsNotSupported;
 import com.oracle.svm.core.jdk.ContinuationsSupported;
 import com.oracle.svm.core.jdk.JDK11OrEarlier;
@@ -63,6 +63,7 @@ import com.oracle.svm.core.jdk.LoomJDK;
 import com.oracle.svm.core.jdk.NotLoomJDK;
 import com.oracle.svm.core.monitor.MonitorSupport;
 import com.oracle.svm.core.util.VMError;
+import com.oracle.svm.util.ReflectionUtil;
 
 @TargetClass(Thread.class)
 @SuppressWarnings({"unused"})
@@ -635,17 +636,27 @@ public final class Target_java_lang_Thread {
     @TargetElement(onlyWith = LoomJDK.class)
     public static native Target_java_lang_Thread_Builder ofVirtual();
 
-    /** This method being reachable fails the image build, see {@link ContinuationsFeature}. */
     @Substitute
     @TargetElement(name = "ofVirtual", onlyWith = {JDK19OrLater.class, NotLoomJDK.class})
     public static Target_java_lang_Thread_Builder ofVirtualWithoutLoom() {
+        if (Target_jdk_internal_misc_PreviewFeatures.isEnabled()) {
+            if (DeoptimizationSupport.enabled()) {
+                throw new UnsupportedOperationException("Virtual threads are not supported together with Truffle JIT compilation.");
+            }
+            if (SubstrateOptions.useLLVMBackend()) {
+                throw new UnsupportedOperationException("Virtual threads are not supported together with the LLVM backend.");
+            }
+        } else {
+            Target_jdk_internal_misc_PreviewFeatures.ensureEnabled(); // throws
+        }
         throw VMError.shouldNotReachHere();
     }
 
-    /** This method being reachable fails the image build, see {@link ContinuationsFeature}. */
     @Substitute
-    @TargetElement(onlyWith = {JDK19OrLater.class, NotLoomJDK.class})
-    static Thread startVirtualThread(Runnable task) {
+    @TargetElement(name = "startVirtualThread", onlyWith = {JDK19OrLater.class, NotLoomJDK.class})
+    static Thread startVirtualThreadWithoutLoom(Runnable task) {
+        Objects.requireNonNull(task);
+        ofVirtualWithoutLoom(); // throws
         throw VMError.shouldNotReachHere();
     }
 
@@ -782,4 +793,13 @@ final class Target_java_lang_Thread_ThreadIdentifiers {
 interface Target_sun_nio_ch_Interruptible {
     @Alias
     void interrupt(Thread t);
+}
+
+@TargetClass(className = "jdk.internal.misc.PreviewFeatures", onlyWith = JDK19OrLater.class)
+final class Target_jdk_internal_misc_PreviewFeatures {
+    @Alias
+    static native boolean isEnabled();
+
+    @Alias
+    static native void ensureEnabled();
 }