+++ truffle/mx.truffle/suite.py
@@ -798,6 +798,12 @@ suite = {
             "ldlibs" : ["-ldl"],
           },
         },
+        "linux-musl" : {
+          "<others>" : {
+            "cflags" : ["-g", "-Wall", "-Werror"],
+            "ldlibs" : ["-ldl"],
+          },
+        },
         "<others>" : {
           "<others>" : {
             "cflags" : ["-g", "-Wall", "-Werror"],